test(dialog-full-screen): add accessibility tests
