+++ cypress/support/accessibility/a11y-utils.js
@@ -67,6 +67,7 @@ export default (from, end) => {
       !prepareUrl[0].startsWith("detail") &&
       !prepareUrl[0].startsWith("help") &&
       !prepareUrl[0].startsWith("toast") &&
+      !prepareUrl[0].startsWith("dialog-full-screen") &&
       !prepareUrl[0].endsWith("test")
     ) {
       urlList.push([prepareUrl[0], prepareUrl[1]]);
